requirements.txt
pip==23.1.2
Django==4.2
django-admin==2.0.2
django-adminlte3==0.1.6
django-cities-light==3.9.2
django-clearcache==1.2.1
django-import-export==3.2.0
django-six==1.0.5
openpyxl==3.1.2
Pillow==9.5.0
psycopg2==2.9.6
pytest==7.3.1
pytest-cov==4.0.0
selenium==4.9.0
django-formtools==2.4.1
django-crispy-forms==2.0

